fix(Heartbeat): rework heartbeat utility on top of the Mx API [ci skip]
The tests for heartbeat currently only pass because we `startTraceRelay` on the second node. This is not good, and also we generate a lot more network traffic than we need to, so I’ve broken the program up to separately address (a) detecting when an `MxEvent` is interesting, (b) periodically pinging each remote node, and (c) firing an alarm when we haven’t heard from a node within a given time interval. It may well be impossible to test (c) without a usable mock network-transport.